refactor: integration -> internal/integration
