+++ backend/controller/console/console_integration_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/alecthomas/assert/v2"
 
 	pbconsole "github.com/TBD54566975/ftl/backend/protos/xyz/block/ftl/v1/console"
-	in "github.com/TBD54566975/ftl/integration"
+	in "github.com/TBD54566975/ftl/internal/integration"
 )
 
 // GetModules calls console service GetModules and returns the response.
