+++ backend/controller/ingress/ingress_integration_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/alecthomas/assert/v2"
 	"github.com/alecthomas/repr"
 
-	in "github.com/TBD54566975/ftl/integration"
+	in "github.com/TBD54566975/ftl/internal/integration"
 )
 
 func TestHttpIngress(t *testing.T) {
