+++ cmd/ftl/integration_test.go
@@ -10,8 +10,7 @@ import (
 
 	"github.com/alecthomas/assert/v2"
 
-	. "github.com/TBD54566975/ftl/integration"
-	"github.com/TBD54566975/ftl/internal/exec"
+	. "github.com/TBD54566975/ftl/internal/integration"
 	"github.com/TBD54566975/ftl/internal/log"
 )
