+++ go-runtime/encoding/encoding_integration_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/alecthomas/assert/v2"
 
-	in "github.com/TBD54566975/ftl/integration"
+	in "github.com/TBD54566975/ftl/internal/integration"
 )
 
 func TestHttpEncodeOmitempty(t *testing.T) {
